Update index.html
v2.30   On desktop and mobile, move the columns button from where it is now to the white space below the header. Make it inline horizontally with the "Showing x games" text. Make it text instead of a colored pill button, but still clickable and the same functionality it has now. Make it the same font size and color as the Showing x games text. Make it right aligned on right side of screen across from that text. Give it this font awesome icon to the left of its label of how many columns are showing: <i class="fa-solid fa-table-columns"></i>. Make sure functionalities stay the same and things are lined up nicely responsively on all screen sizes, so this new placement is right-aligned across from Showing x games.