Update index.html
v2.35
On desktop only, without changing anything on mobile, I want to restructure a lot of the buttons. I want to have exactly the same sort of hamburger menu with buttons in it as mobile has. So in the top right corner of the page, have the hamburger icon that opens this menu be this icon instead <i class="fa-solid fa-circle-user"></i>, which changes to an X icon in the same spot which closes the menu. Within this menu, have all the same buttons as are in the mobile menu, and remove those buttons from where they are currently on desktop. This will leave only the search bar, Filters button, and Reset button when it's there all centered where they are currently.